fix: use u128 suffix for pool values to match contract types
ClaimModal and OracleResolveModal were passing pool values with u64
suffix, but the contract's claim_winnings and resolve_market transitions
expect u128. This would cause type mismatch errors at execution time.

+++ frontend/src/components/ClaimModal.tsx
@@ -74,8 +74,8 @@ export function ClaimModal({
           inputs: [
             winningRecord.raw,
             winningOutcome ? "true" : "false",
-            `${yesPool}u64`,
-            `${noPool}u64`,
+            `${yesPool}u128`,
+            `${noPool}u128`,
           ],
           fee: 500_000,
         });
